fix: correctly inject environment and properties
The BeanDefinitionRegistryPostProcessor bean needs to have a no-args constructor.
It can inject the environment by implementing EnvironmentAware as well.
Using the environment, you can read Spring Boot properties.

WARNING: reading the filters does not work if they are defined using the GatewayCfg
instead of using Spring Boot properties, as far as I can see. This might be an issue.