test: fail with custom filter
Fail with a custom servlet filter in the chain that throws an exception.

WARNINGs:

* The exception from a filter is currently not handled by the global advice
that returns application/problem+json for errors. Instead, an application/json
response is returned by the REST gateway. The client cannot properly handle this
as an error response, tries to parse it into a TopologyResponse, and fails with that,
swallowing the original exception.
* The test adds filters to the application.properties statically for PoC purposes.
If possible, the test should add them dynamically. Alternatively, the filter registration
needs to read the gateway properties instead of Spring Boot properties as they would contain
dynamically added filters as well.