feat: add intents for message subscription migration
Intents are added for both process and message subscriptions:

Only MIGRATED intent is added to ProcessMessageSubscriptionIntent because the migration command
will be received in the process instance's partition. Therefore, we will not receive a command for
migrating a process message subscription.

Both MIGRATE and MIGRATED intents are added to MessageSubscriptionIntent because the message
subscription can be stored in any partition. Therefore, we will send MIGRATE command to that
partition while migrating the message catch event.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>