feat: write proc-msg-sub migrated event
For mapped message catch events we want to migrate the message
subscription. We'll start with migrating the process message
subscription, by appending a MIGRATED event for it.

We've chosen to track the processMessageSubscriptionsToMigrate as a list
while chosing to unsubscribe it (in the filter), rather than looping
over all message subscriptions related to this scope. The problem with
that would be that we'd have to filter out the unmapped entries again.

Note that this commit is not complete, as we still need to change the
record properties and migrate the message subscription record.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>