feat: distribute msg sub migrate command
To migrate message subscription located in another partition (the subscription partition), we use
distribute command and send MessageSubscriptionIntent.MIGRATE to the related partition. There are
two points to note:

1) The order of records arriving to the other partition via command distribution cannot be
guaranteed. For example, when two commands are distributed after each other, then any of them
could fail and be retried. Meaning that, the later one can arrive first successfully.

This is problematic for migrations because we might have same process instance migrated twice and
the message subscription migration of the second migration could arrive at the related partition
first. Because that reason, we used process message subscription key as a distribution key which
must be unique until the distribution is completed. Then, we will check if there are any pending
distributions of a message subscription and reject following migration until the distribution is
finished.

2) While creating a MessageSubscriptionRecord, we do not set message key and variables because they
are not yet known until the message is correlated.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>