test: add REST job controller tests
Moves the ActivateJobsStub to the common gateway module so it can be used by gRPC and REST tests.
The REST JobControllerTest uses the ActivateJobsStub to create test data.
The ResponseObserver in the test is made reusable and resetable, so tests are independent and
don't serve the response data from previous tests.

Ensures jobs are activated and returned immediately if available.
Ensures rejections are properly handled by the job activation endpoint.
Ensures an empty jobs array is returned when no jobs can be activated.
Ensures the partitions are queried round-robin for activation requests.