refactor: clarify REST API enablement
Clears up REST API beans creation in the distribution module for broker and gateway profiles.
In the gateway profile, the REST API is always up, so the respective beans can always be created.
In the broker profile, the REST API is up by default but can be disabled. The beans are thus only
created if the REST API is enabled (using the custom annotation to hide complexity).

Plus, the related custom annotation receives Javadoc to understand the conditions under which
the REST API can be disabled.