test: increase timeout for reconfigure operations
In production code, we use a higher timeout for these requests
because it requires multiple roundtrips between the members. In
the failed test run, we can see that the operation eventually
completed but the request had failed due to the lower timeout.
To prevent the flakiness of the test, it is better to increase
the timeout similar to the production code.