feat: add ExecuteGetRequest and ExecuteGetResponse
These SBE messages will be consumed and produced by the new
GetApiRequestHandler (Get API of the broker).

I've chosen a simple response type: OK, ERROR. This can differentiate
between a success and an failure. We could consider expanding this to
user errors and system errors, but this is also already covered by the
RejectionType enum.

For the rest, I've kept things similar to the Command and Query APIs:
- we want to request based on partitionId, key, and valueType (for
  example, give me a job with key '2' from partition '3'). Note that
  partitionId is not necessary if the key always encodes the partition,
  but this distinction exists in the other APIs as well
- we want to respond a value dynamically. To decode it, the response
  must contain the valueType again, and a variable data encoding of the
  value itself. We also may want to provide state information
  represented by intent. For errors, we can provide an Error value or
  use the rejectionType and rejectionReason for summaries.