fix: run scheduled tasks at a given timestamp
Checkers were scheduled by the SimpleProcessingScheduleService#runDelayed method
which did not guarantee the exact time of execution, but only the delay for
which the scheduler will wait before triggering the task. Unfortunatelly, the
delay may be extended. A fix is to replace the above-mentioned runDelayed method
with the runAt method which will schedule a task to run at a given timestamp or
exactly after it, which guarantees the same delay every time.