wip: actually respond an entire user task record
This required much more work than I expected, but it works. The broker
can now look up a user task, and respond it to the gateway. The gateway
can decode this response (although there is an assertion fault that I
had to uncomment because I don't want to dig too long into small
issues), and then map the response to something that the gateway can
send to the client. For simplicity, this means that we only see {found:
true} or {found: false} at the client. But with some debugging you can
see that the gateway does receive all relevant info (intent and value).