test: remove static modifier of TestStandaloneBroker
`@TestZeebe` annotation can be used to create an instance of `TestStandaloneBroker` object. When
the field is defined as static, the broker instance is created once for all test in the related
test class (see zeebe/qa/util/src/main/java/io/camunda/zeebe/qa/util/junit/
ZeebeIntegrationExtension.java). Otherwise, the test broker re-created for every test case.

Even if we reset the exporter before executing each test case, the order of accessing
to the exporter is not guaranteed. While running the test in a multi threaded environment
this might lead to unordered access to the exporter (see: ReentrantLock.java) so break the
test isolation. We would like to improve the isolation and re-ceate TestStandaloneBroker
for each case.