test: reject migration when prior msg sub migration is not distributed.
Following scenario is the reason why we should reject migration if msg sub command distribution
from the previous migration is not completed:

Partition 1 migrates a process instance that is subscribed to a message catch event. It starts
distributing the msg-sub migration command to partition 2. But the distribution continues to be
retried because of some random reason. (stuck in distributing state)

Partition 1 migrates that same process instance to a different target again before the previous
message subscription migration distribution completed. It starts distributing the msg-sub migration
command to partition 2 for this target as well. This continues to be retried again.

But the order of these msg-sub migrate distribution commands is not guaranteed. When the scenario
mentioned above occurs, it could lead to data corruption on partition 2 (out of sync with partition
1).

So we should not allow migrating the process instance if it's currently distributing a migration
command (e.g. msg sub migration distribution). As a result of introducing command distribution to
Process Instance Migration, a migration is no longer atomic, because distribution is not.

To avoid migrating the instance during the period that the instance is already migrating, we
need to retrieve pending distributions of command distribution and reject migration if there are any
pending distribution for the given distribution key.

As a result, the test case added with this commit verifies the scenario mentioned above. To
correctly test the behaviour, the partition that the process instance and message subscription is
created should be different. Because the test case pauses processing for the msg sub partition to
mimic distribution failure behaviour. I added a TODO to make process instance partition id
deterministic during process instance creation if possible.