feat: reject migration when prior msg sub migration is not distributed
We need to reject migration if there is an incomplete message subscription migration distribution.
That is the result of introducing command distribution to Process Instance Migration; a
migration is no longer atomic, because distribution is not.

To avoid migrating the instance during the period that the instance is already migrating, we
retrieve pending distributions of command distribution and reject migration if there are any
pending distribution for the given distribution key.

In the migration processor, the check is applied inside unsubscribing from message events. It is
because we loop through all message subscriptions of the current active element, and we need to
reject if there is a pending distribution for any subscriptions exist on the active element at that
moment. Some examples:

Example 1:
First migration: "boundary1" -> "boundary2" (mapped)
Second migration: "boundary2" -> "boundary3" (mapped)
Explanation: since the msg sub migration distribution for event "boundary1" is not completed, it
will still belong to process instance at the time of second migration. As a result, "boundary1"
will be seen as an unmapped msg subscription in the second migration attempt. Still, the second
migration should be rejected because of an incomplete msg sub migration distribution.

Example 2:
First migration: "boundary1" -> "boundary1" (mapped)
Second migration: "boundary1" -> "boundary1" (mapped)
Explanation: in this example, "boundary1" will be seen as a mapped msg sub in the second migration
attempt. But since the msg sub migration distribution for event "boundary1" is not completed, the
second migration should be rejected.