feat: append follow-up cmd if sender and receiver partition is the same
If the sender and the receiver partition is the same while sending message subscription migration
command, we can simply append a follow-up command to the current partition. That way, we prevent an
overhead of sending a migration command through inter partition communication.

To determine the partition id for message subscription we need to use partition count. Then, we
compare it with the current partition id.