feat: add message subscription migrate processor
Add the processor for message subscription migration. `DistributedTypedRecordProcessor` is
implemented in the processor class since this command can be received from other partitions as
well. While processing the command in the same partition we simply need to append the MIGRATED
event to update the state. When processing the command in a different partition, in addition to
updating the state, we need to send acknowledgement back to the partition where the command is
sent from.

Please note that we do not distribute command inside processing a new message subscription
migration command as opposed to other `DistributedTypedRecordProcessor` implementations in the
Engine. This is because msg sub migrate command is used to update msg sub properties stored in a
specific partition's state. Therefore, no point of distributing command to other partitions.