test: add regression test case
The process instance migration feature subscribes to unmapped boundary
events, but does not take into account that activating elements may
encounter an incident before subscribing. In other words, when an
incident is raised during activation, there is no subscription. When
migrating the element instance, it should thus not subscribe to the
boundary event either.

Co-authored-by: Nicola Puppa <nicola.puppa@camunda.com>