fix: use millis unit to determine deadline
The timeunit can be nanos (it is on older versions, like 8.2). But we
calculate the deadline as a milliseconds value, and use it as such.

The change doesn't fit the full semantics of the class, as the
constructor's timeunit is not used for the deadline. The deadline is
always in milliseconds. We think that's fine for now.

Co-authored-by: Lena Schönburg <lena@fairburg.xyz>