docs: clarify special case
We wanted to make it a bit more clear why we shouldn't subscribe to new
catch events for ACTIVATING element instances.

For completeness, 19212 popped up because we subscribed an activating
element instance to a message boundary event during migration. It was
activating because there was an incident due to input mappings raised on
it. When resolving the incident, the element instance continued
activating and attempts to subscribe to its catch events. This failed
because migration had wrongly already subscribed the element instance to
it.

To further clarify what is done in this if block, I've changed the
method name and gave the method some javadoc.