fix: copy process msg sub object retrieved from DB
In state classes we use a ProcessMessageSubscription singleton. We put retrieved subscriptions into
a map for migrating them later in the flow. Making a copy of it is required to prevent pointing to
a wrong subscription as we traverse through subscriptions. This is especially required for active
elements with multiple boundary events migrated.