feat: initialize ExporterState based on configured exporters
On first update to the version that added exporters to the dynamic config,
the exporter state will be empty. Also, when a broker is restarted with
new exporters configured or existing exporters removed, this is not reflected
in the dynamic config.

In both of these cases, we want to ensure that the static configuration is reflected
in the dynamic configuration. On startup, after initializing the configuration
each broker checks its dynamic config and the static config for exporters. Any newly
added exporters will be added to the dynamic config with state enabled. If any existing
exporters are removed, they will be marked as disabled.