feat: add appliers for subscription migrations
Appliers are created for message subscription migrations. Update methods are introduced in state
classes to simply update records in the state.

Please note that in DbMessageSubscriptionState, the message subscription is retrieved from the
state. Then its correlating state is set to the migrated message subscription. It is to keep
correlation state of the msg sub record if a correlation is started right before the migration.