fix: fix test failures
Since appliers were not in place at the time of writing ProcessInstanceMigrationClusteredTest, it
now requires an updated element id. Which conforms with the actual usage of the migration case.

Removed the rejection test which is not needed since we now support migrating message boundary
events.

A field should not be capitalised if it is not defined as `static`.