feat: map REST job responses to client domain model
Supports transforming REST responses for job activation to the client's domain model.
Job attributes are transformed to default values if not contained in the REST response.

Job variables are stored as a map from string to object (the original REST response) and
as stringified JSON (using jsonMapper to convert that map).

Custom headers are transformed from a string-object map to a string-string map by
stringifying the original map object values using jsonMapper.