feat: allow merging ClusterConfiguration from 8.5 and 8.6 correctly
During rolling update from 8.5 to 8.6, the brokers that first update to
8.6 will initialize their DynamicConfig in it's own MemberState. When
this is gossiped to brokers in 8.5, they cannot read it and simply ignore
the unknown fields in the message. When those brokers later upgrade to
8.6, they will have an uninitialized DynamicConfig in MemberState of other
members. This commit ensures that the config is updated to a valid one
when it receives an update with a valid config.