test: separate client test REST request creation from handling
Provides a clearer interface for the Java client's RestGatewayService by removing the
internal creation of request data objects. The service methods receive and handle the
request objects, the tests take care of creating them.

The RestGatewayService also provides documentation on service methods, naming the URL the
method corresponds to. Plus, the URL is automatically bound to the versioned context path
the HttpClientFactory defines for the whole REST communication of the client.