fix: remove static modifier of TestStandaloneBroker
As explained in PR https://github.com/camunda/camunda/pull/19081: @TestZeebe annotation can be used
to create an instance of TestStandaloneBroker object. When the field is defined as static, the
broker instance is created once for all test in the related test class (see
ZeebeIntegrationExtension.java). Otherwise, the test broker re-created for every test case.

To provide the test isolation for similar classes where flakiness also occurred for the similar
reason, the static modifier of the TestStandaloneBroker is removed. It will force the broker to be
re-created for every test case.