refactor: allow configuring test stream processor
This change is inspired by StreamPlatform.buildStreamProcessor() which
contains a way to configure the streamprocessor via a parameter. The
consumer lambda has access to a method .streamProcessorMode which
can be used to start the stream processor in such a way that it doesn't
process any new commands. We'll need this to stabilize the
ReplayStateRandomizedPropertyTest.

To achieve this, we can first make it possible to configure the stream
processors (plural as the engine rule can be multi-partition) during
start up.

This commit doesn't change any behavior yet. It keeps the default the
same: StreamProcessorMode.PROCESSING.