test: remove redundant state comparison
The test wants to await that processing has stopped. To make this
deterministic, the test adds a small grace period between two checks
that the engine has reached the end of the log stream. For example,
there could be some new commands written to the log by the timer due
date checker to trigger a timer.

While that is useful, there's no need to additionally compare the state
between these steps. We already compare the state inside the test case
after the restart, i.e. the real property tested by this test case.

Therefore, I've removed the additional comparison. This should speed up
the tests somewhat, as the comparison is expensive.