refactor: allow async start of test engine
When starting the engine, it was always using awaitOpening true. This
means that the method waits for the recoveryLatch. If we were to start
the engine in REPLAY mode, this never happens. Which would slow down the
start up to the maximum wait duration on this latch (15 seconds).

Perhaps we can remove this latch altogether, but for now, I just made
this parameter available to the caller of start.