test: stabilize replay state randomized test
By starting the engine in replay mode, we can make this test
deterministic.

Previously, it was possible that when pausing, by chance some
commands would be appended to the log stream. These commands would not
be processed until the engine restarts. As the engine was restarted in
processing mode, it would then process these command producing new state
changes. This would result in test failure, because the state is
different after the restart.

However, the property that this test case tries to verify is not that
after restart (AND PROCESSING) the state is the same as before the
restart. The property actually is:
- After replay, the state is the same as before the restart

Or in other words:
- Replay is deterministic.

This commit changes the way this test works by restarting the engine in
REPLAY mode in order to compare the state with the prior-to-restart
state.

We comparison of states can be done as soon as the replay has finished.
We can check this by comparing the lastProcessedPosition before and
after the restart.

To ensure that the engine does continue regular processing at some
point, we have to continue the engine is PROCESSING mode as well.
There's no way to do that at this time, so we're simply stopping and
restarting the engine once more.

(cherry picked from commit 0aeca3b9f323020c88e62c11c47fc38204376557)