test: add filter it
Add integration tests for REST API filters.
The IT tests mainly cover that custom filters can be
configured and loaded by the Zeebe REST API in the
Gateway. Test coverage for the Query API or for
multi-tenancy is not added since the Query API is not
provided through REST, and multi-tenancy is implemented
through Spring Security in the REST API.