test: Create integration test for JUnit extension
Create a first integration test for the JUnit extension to verify that the extension works with the real Camunda Docker
image.

Enable the Maven failsafe plugin to run the integration test. By naming the test class with `IT`, it is marked as an
integration test and ignored by the Maven surefire plugin.