feat: add job services for Spring applications
The camunda services configuration for Spring contains a bean for job services
based on the REST data classes. The job services are typed to a specific data type,
handled by the job activation handler and consumed by the response observer that is
passed on to the job services' "activateJobs" method.