feat: remove `ExecutionListenerEventType` enum and consolidate into `ListenerEventType`
Removed the `ExecutionListenerEventType` enum and consolidated event types into the `ListenerEventType` enum. This
simplifies the codebase by having a single enum for event types, which will be used for both Execution Listeners and
User Task Listeners.

Property of `ExecutionListenerEventType` type was not exported before and was only used internally. This change does not
impact any public interfaces or exported APIs.