test: mapped boundary event must stay attached
It's important for instance migration that any boundary event that an
element instance is subscribed to continues to be attached to the same
element instance.

Users could provide mapping instructions tearing the boundary event away
from the element instance it is attached to. This would lead to
processing errors, so we should not allow this.