refactor: fallback to default advertised host if no host specified
In order to distinguish between a host specified explicitly by the user
and using the default host, there is no default value to the host field
of the `NetworkCfg`. Instead, the default host is returned in the getter
if the `host` field is null.

For advertised host, we then use the following logic in the getter:
- Return the `advertisedHost` property if explicitly set
- Return the `host` property if explicitly set
- Return the default advertised host for the machine if nothing is set

This allows us to have a default binding host of `0.0.0.0` (or `::` if
using IPv6), but an appropriate resolvable default advertised host based
on the heuristics from `Address#defaultAdvertisedHost` which should be
resolvable most of the time.

Doing lets us avoid having to explicitly set the `host`/`advertisedHost`
properties in the `Dockerfile` down the line for a better user
experience.

Note that we don't need to touch `CommandApiCfg` and `InternalApiCfg`
because they both will either use explicit values or default to whatever
the `NetworkCfg` returns (so using the heuristics above).