refactor: remove explicit environment configuration
Removes explicit configuration of network hosts which lead to a
confusing UX. Since we set them here in the entry point, the only way
for a user to reconfigure them is also through environment variables -
they cannot be reconfigured via configuration file or system properties.
This is can lead to confusion for users since it's not obvious that only
these settings work this way.

In the future, there should be no configuration which is explicitly set
directly except through the default configuration file.