feat: update subscribeToEvents to filter out events beforehand
A filter added to `CatchEventBehavior.subscribeToEvents` method to filter out events of an element
instance before executing any expression evaluations. This is used to prevent evaluating
expressions on events that are migrated. Because, we are currently not supporting expression
evaluation on migration. This will become optional (e.g. allowing correlation key or message name
re-evaluation) in further migration improvements.