fix: deal with standalone and embedded gateway
The gateway can be embedded in a broker or run standalone. Different
configs are available for this: zeebe.gateway.* and
zeebe.broker.gateway.*.

We can check whether the security.authentication.mode has been disabled
explicitly for either, and in that case disable the gateway security.
This inverse check (disable on one of property) can be achieved using
the NoneNestedConditions.

This config means that the gateway security is enabled:
- is enabled by default
- is enabled when security.authentication.mode is set to identity
  explicitly
- is disabled when security.authentication.mode is set to none
  explicitly

(cherry picked from commit a120b4b771e395833f8bbd2a06bc1a20f5287455)