feat: set binding type attribute via BPMN model builder classes
This change allows to set the binding type in the respective extension element for call activities, business rule tasks,
and user tasks.

Note: In AbstractUserTaskBuilder, the setter is called 'zeebeFormBindingType' instead of just 'zeebeBindingType'.
This should make it clearer that the binding type refers to the form definition,
given that there are a couple of other attributes that are not related to forms at all.