feat: implement timer event migration
Timer event is added to allowed boundary events. A method introduced to CatchEventBehavior to be
able to filter timer events by timer instance. It is needed to keep track of timer instances to
migrate. Please note that we detect migrated timer events by checking if the retrieved timer
instance's process definition key is the target's process definition key. That means, the timer
instance already subscribed for the target process. Lastly, we go over timer instances to migrate
and write MIGRATED event to update them in the state.