refactor: simplify user task update REST data model
The user task update REST endpoint uses a data model that nests another data model.
The nested data model reference is consumed as an `allOf`, indicating that we extend
the original reference. However, we don't actually extend it execpt for adding more
description on top and making it nullable.

This can be straightened out to moving the description and nullable attributes to the
referenced data model directly and using it simply as an unaltered reference.