fix: allow deployment of multiple resources with same name
During the deployment distribution we checked if we had the right resource by resource name, and would send a Process
.CREATED event accordingly. This is not entirely accurate. Users could deploy multiple processes with the same resource
name. This would result in multiple Process.CREATEED event getting appended.

This fix makes it so we use the checksum instead of the resource name to find the resource by the metadata. This way
actually compare the process itself and guarantee it matches the process the metadata belongs to.

(cherry picked from commit 267428ed677d37ed4e5ba52c2e098dd6e1e1c19d)