fix: locklessly replace audiences in cache
The audiences hashmap was not accessed threadsafe. Simply marking the
methods synchronized does not achieve this if the methods themselves can
be called independently from other threads.

We can overcome this by locklessly replacing the hashmap. To do so, we
can use an AtomicReference and introduce a utility replace method that
attempts to replace the reference but only does so if another thread
didn't replace it successfully in the meantime. If it was replaced by
another thread, it will retry replacing it.

This only works as long as the referenced map is not modified while
referenced by AtomicReference.

One downside to this approach is that we need to copy the entire hashmap
everytime we want to put a new value.

(cherry picked from commit 3ff56044b2c5f49ed1f143a39a1e94d0106f6144)