fix: only fetch credentials once at a time
By removing the synchronized previously, we broke the test case:
`OAuthCredentialsProviderTest.shouldCallOauthServerOnlyOnceInMultithreadMode`

This says that we should only make a single Oauth request at a time.
Previously, this was controlled by using synchronized on all cache
methods. But we only need a synchronized on the
`computeIfMissingOrInvalid` method.

This works, because it calls `OAuthCredentialsProvider.fetchCredentials`
as the `zeebeClientCredentialsConsumer`. By blocking on the entire
cache, we prevent the credentials to be fetched multiple times.

This is a crude solution, as different requests to fetch credentials are
likely still needed when different endpoint are used, but the entire
method is locked regardless of the endpoint.

As it's unclear what the performance gain would be from locking per
endpoint, and this previously hasn't been any issue, I'll revert back to
using synchronized on the `computeIfMissingOrInvalid` method.

(cherry picked from commit af15bf2a300356fb024c67b44ab9a31223a948a0)