test: fix flaky raft heartbeat test
After the follower is disconnected, it will send two poll requests.
Waiting for more than 2 poll requests is flaky. The test most of the
time passes because, there is enough time for two rounds of poll requests.
Or sometimes the interceptor is added before the first election is completed.
So the initial poll requests before the follower is disconnected is
counted. To fix it now, we wait until a leader is elected, and wait until
only 2 poll requests are observed.

(cherry picked from commit 9ef7c3a83e169bcf5a4b7f01804abac359d2215e)