fix: ensure client requests retried at most twice
There was a bug where client requests could be retried indefinitely if
the request was responded to with UNAUTHORIZED (HTTP 401 or grpc
UNAUTHENTICATED).

This endless retry could lead to 429 Too Many Requests, at which point
the retry stopped. However, there's no need to retry UNAUTHORIZED
requests so often.

This commit limits the number of retries to two. This means requests are
tried a total of three times.