feat: create MessageCorrelation intents
The MessageCorrelation ValueType will have 3 intents: CORRELATE, CORRELATED and NOT_CORRELATED.

CORRELATE is the command that is sent by the client.
CORRELATED will be used as a response to the client and indicates we managed to correlate the message to a subscription.
NOT_CORRELATED will be used as a response to the client and indicates we did not manage to correlate the message to a
subscription.