feat: add request properties to MessageSubscriptionRecord
In order to be able to send a response back to the client upon correlation we must keep track of the request data. This
is tricky considering multiple partitions are involved in message correlation.
Putting the request properties on this record allows us to store them in our state on the MessageSubscription
.CORRELATING event applier. This way, once we get feedback from the other partition that correlation succeeded, we can
pull these properties from the state in the Messagesubscription.CORRELATED` command. At this point we can use them to
send the response to the client.