fix: unblock controlled actor thread when closing
When a controlled actor thread is closed while idle, we need to unblock
it so that it can wrap up instead of being stuck in idle forever.

This would also be solvable by requiring the caller that wants to stop
the thread to call `resumeTasks`. But doing this in the `close` and
deactivating the blocking idle (by deregistering from the phaser) is
more reliable.