fix: ensure controlled actor scheduler is stopped after tests
Previously, we did not await on the closing future. One reason was that
closing would not work reliably, fixed in dc709e66a.

With this change, we wait for up to 5 seconds before giving up. The time
limit is to ensure that blocked actor tasks don't block all further
tests.