feat: add deployment key to deployment record
Generate the deployment key before the resource transformation is happening and add it to the deployment record,
so that it can be added to process/decision/form records later on.

With the introduction of the new "deployment" binding type, the deployment key now serves a purpose and is more than
just a technical key. Storing the deployment key in the record might also be useful in the future if we wanted to allow
users to delete resources by deployment key, rather than by resource key specifically.