fix: terminate element iteratively instead of recursively
We've seen issues where the modification would run into a StackOverflowError due to the recursive nature of the
terminating logic. By using stacks we can solve this problem iteratively and prevent running out of stack size.

This commit shows the difference between the recursive and iterative method. In the next commit I will clean up the code
 a bit more by refactoring it.