fix: add a read and write lock to the cache file
We were reading and writing to the cache file simultaneously. As a result this wasn't thread safe. By adding a read
write lock we can ensure that we are only reading when there's no other thread writing. And we will only start writing
when there's no other threads reading.
Due to the nature of the ReentrantReadWriteLock multiple threads are allowed to read simultaneously.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>