test: should trigger timer event of a banned instance only once
A test case is created to verify banned instance's timer event is triggered only once. An
integration test chosen to utilize the actual command cache.

Firstly, a process instance is created with timer duration of 2 seconds and the instance is banned.
Later, a process with timer duration of 3 seconds is created to make sure banned instance's timer
is triggered. Lastly, a process with a second duration timer is created to make sure the banned
instance's timer is not triggered again.