fix: allow setting skipped status to processing result
A boolean field is introduced to ProcessingResult and ProcessingResultBuilder to allow updating the
field. Later, this field will be set to `true` when a command is skipped (e.g. banned instance's
commands). When a ProcessingResult is marked as `skipped`, we will not remove the command from the
command cache to prevent appending it to the log over and over again.