feat: use elementId hash in CallingElementPath
Since ElementIds are just a limited number of strings unique locally to the deployed model
we could use a hashing function to hash these Ids in the callingElementTreePath

Operate can use the same hashing function to hash all elementIds in a model and build a reverse `Map<Hash, ElementId>`
and use this map to look up the elementId string using the hash.

I decided to use `murmur3_128` because it is a fast hashing function with low chance of collision.
And I choose an implementation from `google guava`