feat: send response upon message subscription correlation
During processing of the MessageSubscription CORRELATE command we know for sure that the message is correlated. Here we
can write the response back to the client. We do this by fetching the subscription from the state, in which we stored
the request data. We also need to write MessageCorrelation CORRELATED event to complete the event lifecycle.