test: verify message correlation multi-tenancy
I've made this test parameterized as MT works the same for the message publish and the message correlate command. The
test verify that correlation works as expected when a certain tenantId is passed along.

One of the tests verifies message buffering. This is not supported for the message correlate command, and as a result I
am skipping this test for this scenario.