test: reset recordingexporter before test setup
Previously, any IT annotated with `ZeebeIntegration` would have the
recording exporter reset between the `@BeforeEach` and the `@Test`
methods.

If the user, sets up any things that produce records on the log stream,
then we're typically interested in those records in the test case as
well. So, the recording exporter should only be reset before the setup,
i.e. before each.

This was exactly what is wrong in #17668:
- Before the test is run, there is some setup
- This setup uses the RecordingExporter to verify that the set up
  succeeded
- The RecordingExporter has not yet been reset (it will be reset after
  the test setup and before the test is executed)
- This caused records from a test from a previous test class to still
  exist on the RecordingExporter (note that the RecordingExporter is
  static)
- We expect that these unrelated records interfered with the
  assertions in the test setup

Co-authored-by: ana.vinogradova <ana.vinogradova@camunda.com>
Co-authored-by: Mustafa Dagher <mustafa.dagher@camunda.com>
Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Nikola Koevski <nikola.koevski@camunda.com>