feat: resolve called process ID for binding type "deployment"
If binding type "deployment" is used for the called process ID in a call activity, the process ID will be resolved to
the particular version of the child process that was deployed together with the current version of the parent process
(instead of just the latest available version).

If the target process cannot be found in the same deployment, an incident will be raised.

(cherry picked from commit a2b5738afe33f75281325b172be0f17ca0d524a6)