fix: pass empty object as delete action to ES
It's not clear anymore why, but we passed an ObjectMapper as part of a
DTO to ES. The value in json format should be an empty object `{ }` so
passing an ObjectMapper may have simply been a typo which should've been
new Object() instead.

Instead of using new Object(), it's even better to use a record. We can
even make that part of the API to avoid this mistake altogether.

Note that the API is very specific to the Delete action for the Delete
phase only. I did not want to make this more flexible now, as we have no
need for it now. We can do this later if we ever see a need.

(cherry picked from commit 02e8eea0e99afdf947ac0634beff6bd3fd5f259a)