feat: resolve called decision ID for binding type "deployment"
If binding type "deployment" is used for the called decision ID in a business rule task, the decision ID will be
resolved to the particular version of the decision that was deployed together with the current version of the process
(instead of just the latest available version).

If the target decision cannot be found in the same deployment, an incident will be raised.

(cherry picked from commit 1342272e82fb527dc48949ddb305f51411e0d278)