feat: migrate intermediate message catch event
The migration class is updated to support migrating intermediate message catch events.
Preconditions are called to prevent migration of intermediate catch events other than message.
Migrating message subscriptions is refactored to be used by both message boundary events and
message catch events.

Please note that there is a return statement at the end of handling message catch event. It is
because intermediate events cannot have boundary events and if `handleBoundaryCatchEvents` is
called for the intermediate catch event, it fails with type casting errors.