refactor: align REST controller annotation
Renames the ZeebeRestController to CamundaRestController to align on naming as the controllers
are not Zeebe-specific anymore.
Removes the path from the controller annotation to avoid overriding paths in the controllers
which leads to warnings logged by Spring.

(cherry picked from commit e750217c8dda922bf6343377fa46a71f6a25140c)