refactor: activate REST resources consistently
All REST API-related Spring resources must be based on the ConditionalOnRestGatewayEnabled
annotation since the condition it's based on is different for broker and gateway setups,
basing on different properties to be enabled or disabled.

Otherwise, disabling the zeebe.broker.gateway.enable property also disables REST resources in the
gateway use case where it does not make sense (it only watches camunda.rest.enabled).