feat: handle REST query API availability
Supports configuring if the REST Query API is enabled, seperately from the general REST API.
This allows to disable the Query API while the REST API in general is enabled. This is also the
default behavior if no configuration is provided.

The Query API cannot be enabled while the REST API in general is disabled.